requirements.txt
Appium-Python-Client==0.49
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
peewee==3.13.1
python-dotenv==0.11.0
requests==2.22.0
selenium==3.141.0
urllib3==1.25.8
Werkzeug==1.0.0
